requirements.txt
faster-whisper>=1.2.0
sounddevice>=0.4.0
numpy>=1.24.0
evdev>=1.7.0
python-dotenv>=1.0.0
